Refactor mixed terminology of 'provider' and 'client'
