+++ lib/storage/clients/aws.rb
