+++ lib/storage/clients/azure.rb
